Added Loading circle effect/component
