+++ src/components/LoadingCircle/LoadingCircle.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+import "./loading-circle.scss";
+
+const LoadingCircle = () => {
+  return (
+    <div className="spinner">
+      <div class="loadingio-spinner-rolling-rxln3jp17tb">
+        <div class="ldio-3hg0u7kjo9z">
+          <div></div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default LoadingCircle;
