+++ src/components/LoadingCircle/index.js
@@ -1,0 +1,2 @@
+import LoadingCircle from "./LoadingCircle";
+export { LoadingCircle };
