+++ src/components/LoadingCircle/loading-circle.scss
@@ -1,0 +1,47 @@
+@keyframes ldio-3hg0u7kjo9z {
+  0% {
+    transform: translate(-50%, -50%) rotate(0deg);
+  }
+  100% {
+    transform: translate(-50%, -50%) rotate(360deg);
+  }
+}
+.spinner {
+  height: 100%;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  .ldio-3hg0u7kjo9z div {
+    position: absolute;
+    width: 120px;
+    height: 120px;
+    border: 20px solid #c2bbf0;
+    border-top-color: transparent;
+    border-radius: 50%;
+  }
+  .ldio-3hg0u7kjo9z div {
+    animation: ldio-3hg0u7kjo9z 1s linear infinite;
+    top: 100px;
+    left: 100px;
+  }
+  .loadingio-spinner-rolling-rxln3jp17tb {
+    width: 200px;
+    height: 200px;
+    display: inline-block;
+    overflow: hidden;
+    background: transparent;
+  }
+  .ldio-3hg0u7kjo9z {
+    width: 100%;
+    height: 100%;
+    position: relative;
+    transform: translateZ(0) scale(1);
+    backface-visibility: hidden;
+    transform-origin: 0 0; /* see note above */
+  }
+  .ldio-3hg0u7kjo9z div {
+    box-sizing: content-box;
+  }
+  /* generated by https://loading.io/ */
+}
