+++ src/features/Login/login.scss
@@ -1,4 +1,4 @@
-.login {
+.login-container {
   height: 100vh;
   width: 100vw;
   display: flex;
@@ -6,58 +6,4 @@
   justify-content: center;
   align-items: center;
   background-color: rgba(241, 227, 243);
-
-  &__container {
-    height: 40%;
-    width: 60%;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    border: 2px solid rgb(53, 144, 243);
-    border-radius: 20px;
-    background-color: rgba(98, 191, 237);
-    overflow: hidden;
-    box-shadow: 5px 5px 12px 0px rgba(0, 0, 0, 0.3);
-  }
-  &__title {
-    height: 15%;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: rgba(53, 144, 243);
-  }
-  &__form {
-    font-size: 20px;
-    height: 85%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-    align-items: center;
-  }
-  &__mb-3 {
-    height: 33%;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-evenly;
-  }
-  &__input {
-    min-height: 50%;
-    border-radius: 5px;
-    border: none;
-    background-color: rgba(241, 227, 243);
-  }
-  &__button {
-    border: none;
-    border-radius: 5px;
-    height: 30px;
-    width: 10%;
-    background-color: rgba(241, 227, 243);
-    &:hover,
-    &:focus {
-      cursor: pointer;
-      background-color: rgba(194, 187, 240);
-    }
-  }
 }
